requirements.txt
aiofile==1.5.2
aiofiles==0.4.0
aniso8601==7.0.0
click==7.1.1
dataclasses==0.7
envparse==0.2.0
fastapi==0.52.0
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
Jinja2==2.11.1
MarkupSafe==1.1.1
motor==2.1.0
promise==2.3
pydantic==1.4
PyJWT==1.7.1
pymongo==3.10.1
python-multipart==0.0.5
Rx==1.6.1
six==1.14.0
starlette==0.13.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
